refactor: Participants를 상속함으로써 Dealer와 Player의 중복 코드를 제거
