+++ src/main/java/domain/Dealer.java
@@ -1,36 +1,13 @@
 package domain;
 
-import java.util.List;
-
-public class Dealer {
+public class Dealer extends Participants {
     private static final String DEALER_NAME = "딜러";
-    private final Hand hand;
 
     public Dealer(Hand hand) {
-        this.hand = hand;
-    }
-
-    public void keepCard(Card card) {
-        hand.addCard(card);
-    }
-
-    public Integer handSize() {
-        return hand.getHandSize();
-    }
-
-    public Integer getTotalCardScore() {
-        return hand.calculateTotalScore();
+        super(DEALER_NAME, hand);
     }
 
     public Boolean dealerRule() {
         return getTotalCardScore() < 17;
-    }
-
-    public String getName() {
-        return DEALER_NAME;
-    }
-
-    public Hand getHand() {
-        return hand;
     }
 }
